refactor: Improve formatting of organization_id column migration for better readability

+++ apps/api/database/migrations/18_fix_team_members.ts
@@ -1,4 +1,3 @@
-
 import BaseSchema from '@ioc:Adonis/Lucid/Schema'
 
 export default class extends BaseSchema {
@@ -6,13 +5,21 @@ export default class extends BaseSchema {
 
   public async up() {
     this.schema.alterTable(this.tableName, (table) => {
-      table.integer('organization_id').unsigned().references('id').inTable('organizations').onDelete('CASCADE')
+      table
+        .integer('organization_id')
+        .unsigned()
+        .references('id')
+        .inTable('organizations')
+        .onDelete('CASCADE')
     })
   }
 
   public async down() {
-    this.schema.alterTable(this.tableName, (table) => {
-      table.dropColumn('organization_id')
-    })
+    if (await this.schema.hasTable(this.tableName)) {
+      this.schema.alterTable(this.tableName, (table) => {
+        table.dropForeign(['organization_id'])
+        table.dropColumn('organization_id')
+      })
+    }
   }
 }
